backend/requirements.txt
# ========================================
# Werewolf Arena Backend Dependencies
# ========================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Async & WebSocket
websockets==12.0
aiofiles==23.2.1

# LLM Clients
openai==1.3.0
anthropic==0.7.0

# Configuration
pyyaml==6.0.1

# Utilities
jinja2==3.1.2
tqdm==4.66.1
pandas==2.1.4
psutil==5.9.6
requests==2.31.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Development Tools
black==23.12.0
isort==5.13.0
mypy==1.7.0
ruff==0.1.6

# Legacy dependencies (from original project)
absl-py==2.0.0
marko==2.0.3
